refactor(canvas): extract useCanvasModel so CanvasArea is render-only
